Refactor crypto API to just use PKCS without any extra metadata.
