+++ extensions/amp-a4a/0.1/amp-a4a.js
@@ -13,6 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
+import {
+  LegacySignatureVerifier,
+  PublicKeyInfoDef,
+} from './legacy-signature-verifier';
 import {
   is3pThrottled,
   getAmpAdRenderOutsideViewport,
@@ -39,7 +44,6 @@ import {viewerForDoc} from '../../../src/services';
 import {xhrFor} from '../../../src/services';
 import {endsWith} from '../../../src/string';
 import {platformFor} from '../../../src/services';
-import {cryptoFor} from '../../../src/crypto';
 import {isExperimentOn} from '../../../src/experiments';
 import {setStyle} from '../../../src/style';
 import {assertHttpsUrl} from '../../../src/url';
@@ -121,7 +125,7 @@ let CreativeMetaDataDef;
  * (promises to) keys, in the order given by the return value from the
  * signing service.
  *
- * @typedef {{serviceName: string, keys: !Array<!Promise<!../../../src/crypto.PublicKeyInfoDef>>}}
+ * @typedef {{serviceName: string, keys: !Array<!Promise<!PublicKeyInfoDef>>}}
  */
 let CryptoKeysDef;
 
@@ -249,8 +253,8 @@ export class AmpA4A extends AMP.BaseElement {
     /** @private {boolean} whether creative has been verified as AMP */
     this.isVerifiedAmpCreative_ = false;
 
-    /** @private @const {!../../../src/service/crypto-impl.Crypto} */
-    this.crypto_ = cryptoFor(this.win);
+    /** @private @const {!LegacySignatureVerifier} */
+    this.verifier_ = new LegacySignatureVerifier(this.win);
 
     /** @private {?ArrayBuffer} */
     this.creativeBody_ = null;
@@ -492,7 +496,7 @@ export class AmpA4A extends AMP.BaseElement {
    * @private
    */
   shouldInitializePromiseChain_() {
-    if (!this.crypto_.isCryptoAvailable()) {
+    if (!this.verifier_.isAvailable()) {
       return false;
     }
     const slotRect = this.getIntersectionElementLayoutBox();
@@ -792,7 +796,7 @@ export class AmpA4A extends AMP.BaseElement {
             }
             const signatureVerifyStartTime = this.getNow_();
             // If the key exists, try verifying with it.
-            return this.crypto_.verifySignature(
+            return this.verifier_.verifySignature(
                 new Uint8Array(creative),
                 signature,
                 keyInfo)
@@ -1149,7 +1153,7 @@ export class AmpA4A extends AMP.BaseElement {
    * @private
    */
   getKeyInfoSets_() {
-    if (!this.crypto_.isCryptoAvailable()) {
+    if (!this.verifier_.isAvailable()) {
       return [];
     }
     return this.getSigningServiceNames().map(serviceName => {
@@ -1183,7 +1187,7 @@ export class AmpA4A extends AMP.BaseElement {
               return {
                 serviceName: jwkSet.serviceName,
                 keys: jwkSet.keys.map(jwk =>
-                  this.crypto_.importPublicKey(jwkSet.serviceName, jwk)
+                  this.verifier_.importPublicKey(jwkSet.serviceName, jwk)
                       .catch(err => {
                         user().error(TAG, this.element.getAttribute('type'),
                             `error importing keys for: ${jwkSet.serviceName}`,
