+++ src/service/crypto-impl.js
@@ -14,17 +14,11 @@
  * limitations under the License.
  */
 
-import {PublicKeyInfoDef} from '../crypto';
 import {registerServiceBuilder, getService} from '../service';
 import {dev} from '../log';
 import {extensionsFor} from '../services';
-import {stringToBytes, utf8Encode} from '../utils/bytes';
-import {
-  base64UrlDecodeToBytes,
-  base64UrlEncodeFromBytes} from '../utils/base64';
-
-/** @const {number} */
-const VERSION = 0x00;
+import {stringToBytes, utf8EncodeSync} from '../utils/bytes';
+import {base64UrlEncodeFromBytes} from '../utils/base64';
 
 /** @const {string} */
 const TAG = 'Crypto';
@@ -129,102 +123,68 @@ export class Crypto {
   }
 
   /**
-   * Convert a JSON Web Key object to a browser-native cryptographic key and
-   * compute a hash for it.  The caller must verify that Web Cryptography is
-   * available using isCryptoAvailable before calling this function.
+   * Checks whether Web Cryptography is available, which is required for PKCS 1
+   * operations. SHA-384 operations do not need this because there's a polyfill.
+   * This could be false if the browser does not support Web Cryptography, or if
+   * the current browsing context is not secure (e.g., it's on an insecure HTTP
+   * page, or an HTTPS iframe embedded in an insecure HTTP page).
    *
-   * @param {string} serviceName used to identify the signing service.
-   * @param {!JsonObject} jwk An object which is hopefully an RSA JSON Web Key.  The
-   *     caller should verify that it is an object before calling this function.
-   * @return {!Promise<!PublicKeyInfoDef>}
+   * @return {boolean} whether Web Cryptography is available
    */
-  importPublicKey(serviceName, jwk) {
-    dev().assert(this.isCryptoAvailable());
-    // WebKit wants this as an ArrayBufferView.
-    return (this.isWebkit_ ?
-          utf8Encode(JSON.stringify(jwk)) : Promise.resolve(jwk))
-        .then(encodedJwk => this.subtle_.importKey(
+  isPkcsAvailable() {
+    return Boolean(this.subtle_) && this.win_['isSecureContext'] !== false;
+  }
+
+  /**
+   * Converts an RSA JSON Web Key object to a browser-native cryptographic key.
+   * As a precondition, `isPkcsAvailable()` must be `true`.
+   *
+   * @param {!Object} jwk a deserialized RSA JSON Web Key, as specified in
+   *     Section 6.3 of RFC 7518
+   * @return {!Promise<!webCrypto.CryptoKey>}
+   * @throws {TypeError} if `jwk` is not an RSA JSON Web Key
+   */
+  importPkcsKey(jwk) {
+    dev().assert(this.isPkcsAvailable());
+    return /** @type {!Promise<!webCrypto.CryptoKey>} */ (
+        this.subtle_.importKey(
             'jwk',
-            encodedJwk,
-            {name: 'RSASSA-PKCS1-v1_5', hash: {name: 'SHA-256'}},
-            true,
-            ['verify']))
-        .then(cryptoKey => {
-          // We do the importKey first to allow the browser to check for
-          // an invalid key.  This last check is in case the key is valid
-          // but a different kind.
-          if (typeof jwk['n'] != 'string' || typeof jwk['e'] != 'string') {
-            throw new Error('missing fields in JSON Web Key');
-          }
-          const mod = base64UrlDecodeToBytes(jwk['n']);
-          const pubExp = base64UrlDecodeToBytes(jwk['e']);
-          const lenMod = lenPrefix(mod);
-          const lenPubExp = lenPrefix(pubExp);
-          const data = new Uint8Array(lenMod.length + lenPubExp.length);
-          data.set(lenMod);
-          data.set(lenPubExp, lenMod.length);
-          // The list of RSA public keys are not under attacker's
-          // control, so a collision would not help.
-          return this.subtle_.digest({name: 'SHA-1'}, data)
-              .then(digest => ({
-                serviceName,
-                cryptoKey,
-                // Hash is the first 4 bytes of the SHA-1 digest.
-                hash: new Uint8Array(/** @type {ArrayBuffer} */(digest), 0, 4),
-              }));
-        });
+            // WebKit wants this as an ArrayBufferView.
+            this.isWebkit_ ? utf8EncodeSync(JSON.stringify(jwk)) :
+                             /** @type {!webCrypto.JsonWebKey} */ (jwk),
+            {name: 'RSASSA-PKCS1-v1_5', hash: {name: 'SHA-256'}}, true,
+            ['verify']));
   }
 
   /**
-   * Verifies signature was signed with private key matching public key given.
-   * Does not verify data actually matches signature (use verifySignature).
-   * @param {!Uint8Array} signature the RSA signature.
-   * @param {!PublicKeyInfoDef} publicKeyInfo the RSA public key.
-   * @return {boolean} whether signature was generated using hash.
+   * Verifies an RSASSA-PKCS1-v1_5 signature with a SHA-256 hash. As a
+   * precondition, `isPkcsAvailable()` must be `true`.
+   *
+   * @param {!webCrypto.CryptoKey} key an RSA public key
+   * @param {!Uint8Array} signature an RSASSA-PKCS1-v1_5 signature
+   * @param {!BufferSource} data the data that was signed
+   * @return {!Promise<boolean>} whether the signature is correct for the given
+   *     data and public key
    */
-  verifyHashVersion(signature, publicKeyInfo) {
-    // The signature has the following format:
-    // 1-byte version + 4-byte key hash + raw RSA signature where
-    // the raw RSA signature is computed over (data || 1-byte version).
-    // If the hash doesn't match, don't bother checking this key.
-    return signature.length > 5 && signature[0] == VERSION &&
-        hashesEqual(signature, publicKeyInfo.hash);
-  }
-
-  /**
-   * Verifies RSA signature corresponds to the data, given a public key.
-   * @param {!Uint8Array} data the data that was signed.
-   * @param {!Uint8Array} signature the RSA signature.
-   * @param {!PublicKeyInfoDef} publicKeyInfo the RSA public key.
-   * @return {!Promise<!boolean>} whether the signature is valid for
-   *     the public key.
-   */
-  verifySignature(data, signature, publicKeyInfo) {
-    dev().assert(this.isCryptoAvailable());
-    if (!this.verifyHashVersion(signature, publicKeyInfo)) {
-      return Promise.resolve(false);
-    }
-    // Verify that the data matches the raw RSA signature, using the
-    // public key.
-    // Append the version number to the data.
-    const signedData = new Uint8Array(data.length + 1);
-    signedData.set(data);
-    signedData[data.length] = VERSION;
-
+  verifyPkcs(key, signature, data) {
+    dev().assert(this.isPkcsAvailable());
     return /** @type {!Promise<boolean>} */ (this.subtle_.verify(
-        {name: 'RSASSA-PKCS1-v1_5', hash: {name: 'SHA-256'}},
-        publicKeyInfo.cryptoKey,
-        signature.subarray(5),
-        signedData));
+        {name: 'RSASSA-PKCS1-v1_5', hash: {name: 'SHA-256'}}, key, signature,
+        data));
   }
 
   /**
-   * Is this service actually available? For now, we use browser native
-   * crypto. So if that is not available, then this service is not available.
-   * @return {boolean}
+   * Returns the SHA-1 hash of the input array in a number array. As a
+   * precondition, `isPkcsAvailable()` must be `true` (there's no polyfill
+   * because only Fast Fetch uses this).
+   *
+   * @param {!Uint8Array} input
+   * @return {!Promise<!ArrayBuffer>}
    */
-  isCryptoAvailable() {
-    return !!this.subtle_;
+  sha1(input) {
+    dev().assert(this.isPkcsAvailable());
+    return /** @type {!Promise<!ArrayBuffer>} */ (
+        this.subtle_.digest({name: 'SHA-1'}, input));
   }
 }
 
@@ -233,41 +193,6 @@ function getSubtle(win) {
     return null;
   }
   return win.crypto.subtle || win.crypto.webkitSubtle || null;
-}
-
-/**
- * Appends 4-byte endian data's length to the data itself.
- * @param {!Uint8Array} data
- * @return {!Uint8Array} the prepended 4-byte endian data's length together with
- *     the data itself.
- */
-function lenPrefix(data) {
-  const res = new Uint8Array(4 + data.length);
-  res[0] = (data.length >> 24) & 0xff;
-  res[1] = (data.length >> 16) & 0xff;
-  res[2] = (data.length >> 8) & 0xff;
-  res[3] = data.length & 0xff;
-  res.set(data, 4);
-  return res;
-}
-
-/**
- * Compare the hash field of the signature to keyHash.
- * Note that signature has a one-byte version, followed by 4-byte hash.
- * @param {?Uint8Array} signature
- * @param {?Uint8Array} keyHash
- * @return {boolean} signature[1..5] == keyHash
- */
-function hashesEqual(signature, keyHash) {
-  if (!signature || !keyHash) {
-    return false;
-  }
-  for (let i = 0; i < 4; i++) {
-    if (signature[i + 1] !== keyHash[i]) {
-      return false;
-    }
-  }
-  return true;
 }
 
 /**
