+++ test/functional/testdata-experiments.js
@@ -1,0 +1,8 @@
+/* eslint-disable */
+// Generated from generate-testdata-experiments.js_.
+export const publicJwk = {"alg":"RS256","e":"AQAB","ext":true,"key_ops":["verify"],"kty":"RSA","n":"o9TRgIWUfKNbtdiuLIESc78RRix6wZjYM4pqkRXTDfeersCHQm_zkeOfhrlk6_7dxhd2Oi2zpeKDpG3ERXsRUvL7M39HT9ZfNDiEdMbxNn_3A3NdV3Lyu94iQ8pGtMiv6qmlILIG72zgVlyYFqYryI1wKrgkQER_GnNZf69lRc4EaB3pbPRq-uVR7aP98ysQz6JJ4JvcLJUF_qAVCqnDYUAuRy-Bl77s06f5i4HssHnZgG1nGv3BTkvwT5hTi3BefBvgpy0Nxiu-g3F3pkWAAibV51AUaK-RF9vKoC3-CNzfZ1bHEE3Y1oB4-4jzEo11gK0Bp1pBRVgMAQmwFVbEDQ"};
+export const correctToken = 'AAAAAGB7Im9yaWdpbiI6Imh0dHBzOi8vd3d3Lmdvb2dsZS5jb20iLCJleHBlcmltZW50IjoiYW1wLWV4cGlyZXMtbGF0ZXIiLCJleHBpcmF0aW9uIjo5NTYxNzYwMjAwMDAwMH0J4w8f0OG7WcBcm+qiol0BobuCP8Cremf8koTftyrHNLKN8nmmQ/OPxjsQr8JMCPKSBoVsyZ5LOh+aHo7tFqtaVPjv4wke7U3dUKVIwaxgZyRoO8z5nIzyCSupdW8XNUCkpJMQSLIS0PKBMvWyxZ/D+cKZt0ikrFiuQzYI93zxpFdvg8GOy6is5DfxYiIiz9HZdmvA1cE03HZH6Cau7UR10afkitDkrJrhjVO+mrujCbqpW/Tv+b918M8mrXYKUawKBzn8QELw6gnReE9gw/hh2tU+9LNIWv4i0Kpo4OE0yh/VkDubWYrDGHP4Z5E3QQG6M7s2m1vKDSBUbzRIztzl';
+export const tokenWithBadVersion = 'LwAAAGB7Im9yaWdpbiI6Imh0dHBzOi8vd3d3Lmdvb2dsZS5jb20iLCJleHBlcmltZW50IjoiYW1wLWV4cGlyZXMtbGF0ZXIiLCJleHBpcmF0aW9uIjo5NTYxNzYwMjAwMDAwMH1jZwmgOhi6dl5Tb0GqKKsQ3gz7y3Numqd5wNvoRMZIqWhA8FXBGeP97tj4ZohKLXxSiGkn6IkSiR/3XL+n/vkwmq0uxpSxmkZKznjKNSsu2vwTvbV098UTouU0xi6Z9SP7brm2Bu5n1L5j/i6U+wnZk6LgapK1hejM/i8IkWZ0rnfv+NCreG9jbVTpHNIfk4O+X9v7BQirY1uZblzJuybXDganno6ufGmuXmw763LR1UykY9O2EfheHPV7th9jjbDWpcuJl1O5XAJLcgUZ9FLk/K3luiYKRgbpc9Eqkyzze12JgOE7vmArimmo6oLS06RxkawopZ1I+DhLzUTiQ2Y+';
+export const tokenWithBadConfigLength = 'AAAPQj97Im9yaWdpbiI6Imh0dHBzOi8vd3d3Lmdvb2dsZS5jb20iLCJleHBlcmltZW50IjoiYW1wLWV4cGlyZXMtbGF0ZXIiLCJleHBpcmF0aW9uIjo5NTYxNzYwMjAwMDAwMH0BtoiXA4aHYQ2FNm+vet7BSiDnbwQ+YJv0+aTGgp21S0H59nGP4BKPQjTyrPGex40JsoPBTew8kGJwsqp+tnJNUMrd7gO6h5G7CkNINbkMrAn12KLVlp4BERmJHL1qW2StYLbQrtxKe+tSDuBcgXK8ctC+QOSjo9d1ngm+EwcmTYQKIRUhtCz+8iPIAVc3bR5ipWmEb49D0nIVKKcEe4IN6lpiQdY8ig29HMrnJTMlRW23XeX6tHVWf4utY6KrWy4wCOVPOQ3uBQ9qRs+ECreDBOcHLvKLY6x4EanlfWJHTSIg++DopFJJzPDgMEin4DunS7G0N16FQ+iI52iRWcpP';
+export const tokenWithBadSignature = 'AAAAAGB7Im9yaWdpbiI6Imh0dHBzOi8vd3d3Lmdvb2dsZS5jb20iLCJleHBlcmltZW50IjoiYW1wLWV4cGlyZXMtbGF0ZXIiLCJleHBpcmF0aW9uIjo5NTYxNzYwMjAwMDAwMH0LCGrdK+VBeGBe6WiuwoYmCBYbIdC3XwL2HpKBpQJRdAlypztctfPEb9IMvtfEdIuK1v7A1Ze0Op2nFAttZ0sqwujeQGQQBC7+S5MDafSo4fLZrsKoy0bRWa/bqDI+do9ZFQfNeGdI2NuXJFrJyi1LCm/H1YsCBRSbxDRL7ux2k1ImAOBj+KARLkhgJmXy2iBq61jnhzd3eCzGIAeAVY6aFmV4LHeVat6MC7cbSerjmhpwCJHMiJLzd8vvrpga08fvP4D6oE5WOSNb79Z8844vY22MojucSr/6WmBY4rLHlrBnZ9qZRUWMwcncb0BvlfWyfK1AYNt58+BLi81Nej4R';
+export const tokenWithExpiredExperiment = 'AAAAAFl7Im9yaWdpbiI6Imh0dHBzOi8vd3d3Lmdvb2dsZS5jb20iLCJleHBlcmltZW50IjoiYW1wLWV4cGlyZWQiLCJleHBpcmF0aW9uIjoxMjMyNDI3NjAwMDAwfR3iQl5HEBD6xflmRSRlk/sRngpC07rrsAWvA8IHq3UM95834d3oheHi3P2VbQofxWCNxzEhHotz6qCPl5GS6lH51iy0QSbe3XZrJaOcQQuZfYF4UVqK29PViIYvTRYDj0iqZhQSD1HFTytwD7UY2+pLzJwbgoiI+ENSTeOtzhxl20x7hD+4usxMeE5m18G8r90vkbPfMrqNG4xg34zCox1f3aK7SVgOU8s874c/77ozcfaaIAnL4P/Qyy/1K661DfOleOH79TDFDMCeUVVPI28X3dpUPA7zyAtbgssCzgS9IPGcmsQHC7Pxkde5gL7xD8j1FfrjQ/sT32go8WOUTlo=';
